+++ src/hint_processor/builtin_hint_processor/builtin_hint_processor_definition.rs
@@ -4,13 +4,17 @@ use super::{
         ec_recover_sub_a_b,
     },
     field_arithmetic::uint384_div,
-    secp::ec_utils::{ec_negate_embedded_secp_p, ec_negate_import_secp_p},
     secp::{
-        ec_utils::{compute_slope_and_assing_secp_p, ec_double_assign_new_y},
-        secp_utils::{SECP_P, SECP_P_V2},
+        ec_utils::{
+            compute_slope_and_assing_secp_p, ec_double_assign_new_y, ec_negate_embedded_secp_p,
+            ec_negate_import_secp_p,
+        },
+        secp_utils::{ALPHA, ALPHA_V2, SECP_P, SECP_P_V2},
     },
-    vrf::fq::inv_mod_p_uint256,
-    vrf::{fq::uint512_unsigned_div_rem, inv_mod_p_uint512::inv_mod_p_uint512},
+    vrf::{
+        fq::{inv_mod_p_uint256, uint512_unsigned_div_rem},
+        inv_mod_p_uint512::inv_mod_p_uint512,
+    },
 };
 use crate::hint_processor::builtin_hint_processor::secp::ec_utils::ec_double_assign_new_x;
 use crate::{
@@ -452,12 +456,23 @@ impl HintProcessor for BuiltinHintProcessor {
                 &hint_data.ids_data,
                 &hint_data.ap_tracking,
             ),
-            hint_code::EC_DOUBLE_SCOPE => compute_doubling_slope(
+            hint_code::EC_DOUBLE_SCOPE_V1 => compute_doubling_slope(
                 vm,
                 exec_scopes,
                 &hint_data.ids_data,
                 &hint_data.ap_tracking,
                 "point",
+                &SECP_P,
+                &ALPHA,
+            ),
+            hint_code::EC_DOUBLE_SCOPE_V2 => compute_doubling_slope(
+                vm,
+                exec_scopes,
+                &hint_data.ids_data,
+                &hint_data.ap_tracking,
+                "point",
+                &SECP_P_V2,
+                &ALPHA_V2,
             ),
             hint_code::EC_DOUBLE_SCOPE_WHITELIST => compute_doubling_slope(
                 vm,
@@ -465,6 +480,8 @@ impl HintProcessor for BuiltinHintProcessor {
                 &hint_data.ids_data,
                 &hint_data.ap_tracking,
                 "pt",
+                &SECP_P,
+                &ALPHA,
             ),
             hint_code::COMPUTE_SLOPE_V1 => compute_slope_and_assing_secp_p(
                 vm,