+++ src/tests/cairo_run_test.rs
@@ -858,6 +858,13 @@ fn cairo_run_fq_test() {
 
 #[test]
 #[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
+fn cairo_run_ec_negate() {
+    let program_data = include_bytes!("../../cairo_programs/ec_negate.json");
+    run_program_simple_with_memory_holes(program_data.as_slice(), 0);
+}
+
+#[test]
+#[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
 fn cairo_run_compute_slope_v2_test() {
     let program_data = include_bytes!("../../cairo_programs/compute_slope_v2.json");
     run_program_simple(program_data.as_slice());
@@ -865,7 +872,7 @@ fn cairo_run_compute_slope_v2_test() {
 
 #[test]
 #[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
-fn cairo_run_ec_negate() {
-    let program_data = include_bytes!("../../cairo_programs/ec_negate.json");
-    run_program_simple_with_memory_holes(program_data.as_slice(), 0);
+fn cairo_run_compute_doubling_slope_v2_test() {
+    let program_data = include_bytes!("../../cairo_programs/compute_doubling_slope_v2.json");
+    run_program_simple(program_data.as_slice());
 }